+++ src/main/java/study/racing/domain/Name.java
@@ -1,4 +1,4 @@
-package study.racing.model;
+package study.racing.domain;
 
 public class Name {
