+++ src/main/java/study/racing/domain/car/Car.java
@@ -1,10 +1,10 @@
-package study.racing.model.car;
+package study.racing.domain.car;
 
 import java.util.Objects;
 
-import study.racing.model.Distance;
-import study.racing.model.Name;
-import study.racing.model.rule.Rule;
+import study.racing.domain.Distance;
+import study.racing.domain.Name;
+import study.racing.domain.rule.Rule;
 
 public class Car {
