+++ src/main/java/study/racing/domain/rule/Rule.java
@@ -1,4 +1,4 @@
-package study.racing.model.rule;
+package study.racing.domain.rule;
 
 public interface Rule {
