+++ src/test/java/study/racing/domain/DistanceTest.java
@@ -1,4 +1,4 @@
-package study.racing.model;
+package study.racing.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
