+++ src/test/java/study/racing/domain/car/CarTest.java
@@ -1,4 +1,4 @@
-package study.racing.model.car;
+package study.racing.domain.car;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -9,9 +9,9 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import study.racing.model.Distance;
-import study.racing.model.Name;
-import study.racing.model.rule.Rule;
+import study.racing.domain.Distance;
+import study.racing.domain.Name;
+import study.racing.domain.rule.Rule;
 
 public class CarTest {
